# docs/swagger.yaml
basePath: /
definitions:
  handlers.CreateCollectionRequest:
    properties:
      name:
        example: restaurants
        type: string
    required:
    - name
    type: object
...
  models.Collection:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        description: e.g., "users", "products"
        type: string
      projectId:
        type: integer
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A multi-tenant JSONB document store for students.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: VelocityDBGo API
  version: "1.0"
paths:
...
  /api/projects/{projectId}/collections:
    get:
      description: Retrieve all collections for a specific project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Collection'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all collections
      tags:
      - Collections
    post:
      consumes:
      - application/json
      description: Create a virtual table within a specific project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Collection parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCollectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Collection'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new collection
      tags:
      - Collections
  /api/projects/{projectId}/collections/{collectionName}:
    delete:
      description: Delete an entire collection and all its associated documents
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Collection Name
        in: path
        name: collectionName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a collection
      tags:
      - Collections
...
securityDefinitions:
  ApiKeyAuth:
    description: Public API Key for a specific Project.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"